server_requirements.txt
altgraph==0.17.4
anyio==3.5.0
blinker==1.8.1
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
cloudscraper==1.2.71
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.0
future==1.0.0
GPUtil==1.4.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
icmplib==3.0.4
idna==3.3
impacket==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.3
ldap3==2.9.1
ldapdomaindump==0.9.4
loguru==0.6.0
MarkupSafe==2.1.5
maxminddb==2.6.1
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.4.8
pycryptodomex==3.20.0
pydantic==1.9.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyRoxy @ git+https://github.com/MatrixTM/PyRoxy.git@ea0f88dbc0573292ba5672124f69e4e7a31b544d
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyudev==0.24.1
requests==2.27.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
setuptools==69.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.66.2
typing_extensions==4.2.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.2
yarl==1.9.4
